+++ Brawlmon/Tile.cpp
@@ -1,5 +1,5 @@
 #include "Tile.h"
 
-Tile::Tile(std::string filename, int x, int y,int width,int height) : Actor(filename, x, y, width, height) {
+Tile::Tile(std::string filename, float x, float y,int width,int height) : Actor(filename, x, y, width, height) {
 
 }
